chore(web): extract Enforcement panel to /theguard/policies/enforcement
Phase 2 Chunk A panel 4 of 4 — final panel. Moves the Enforcement tab
out of the Guard settings tabs into a dedicated route +
features/guard/policies/enforcement/Panel.tsx.

With all four panels extracted, /theguard/settings has nothing left to
render. Collapses page.tsx (365 LOC) to a 25-line client-side redirect
that maps every legacy ?tab=* to its new URL and defaults to
/theguard/policies/enforcement. AppShell and GuardShell rail links to
/theguard/settings still work via the redirect — cosmetic label/href
follow-up left for a later PR.
